CSS - Styling the game
First, we're going to throw out any preconceived notions the browser
might have about what we want.  This isn't a document.  This is a game.

Then, we give the board a limiting area denoted by a border, and a
background that separates the board better from the rest of the page.

Then we define what our pieces look like.  They're images, so we'll just
use the existing image tag.  We're going to give them a little spacing
using a margin, and define them to be a standard size.

Finally, we're assigning a font size of 0 to the board itself.  We don't
want empty text nodes messing up our layout.