JS -- Even out the playing field
So we've gotten our game board, but upon loading the page, it contains a
random amount of pieces that already match.

Well.  That isn't fair to all players.  We need to remove these before
the game starts in order start at a score of zero for everybody.

We give the pieces the ability to cycle to the next shape available for
this initial match.

We then tell the board that it is allowed to do an initial matching of
pieces in order to remove possible matches.  We then do the matching by
rows, and then by columns.  Since we can't match past the end of the
playing field, we set a range of areas where it is possible to match.

We then check for the matches at startup.  If a match is found, we
recursively call the match function until no consecutively placed pieces
are found.

But this is just in our model.  We need to reflect this in the
HTML/view.  We go through the board and change all of the images to
match the game piece objects in the model, and then we're ready to start
with the game, so we set it to be the first state -- idle.  Because
nothing is happening right now.