JS/CSS - Give the player feedback
Our game is mouse-powered (perhaps with the help of electrons), so we
add a click handler callback to help drive our finite state machine.

When we click the visual representation of a game piece, we tell the
game that a swap is imminent.  After all, that's the point of clicking.
We then add in a smooth pulsing CSS3 "glow" to indicate that a piece was
clicked.  We also set that this is the initial piece we'll be swapping
so we can reference it later, and an outline as an additional indicator.