JS/CSS - Game completion
We add an animatable CSS property to fade out pieces (for when we get
around to implementing animation) and fix a bug that was causing the
wrong shapes to populate in the view.

We then throw in a loop to continue the match/drop cycle until no more
matches can be made, and reset the state to "idle" to receive more
clicks and further gameplay.