CSS/SQL/PHP/HTML - Cleanup/Base score files
First, we make a small change to the CSS to keep our naming consistent.

Next, we set up a table for holding our score data.  We'll need an
identifier for recording a different score for the same player, a user
column for storing the player name, a score column, and a timestamp (for
posterity).

Then is our PHP file for score handling.  We create a Score class that
can be serialized to JSON (we don't want to refresh in the middle of a
game), and a constructor that uses dependency injection for the database
object (because that's how we roll).

Then we set up the connection information (simple things for a simple
process) and use PDO to connect to the database.  We use PDO because it
allows us to easily change the type of database if our data store ever
changes.

Finally, we add some hooks into the HTML for data display.