CSS/JS/PHP - Get high scores
First, the easy part -- we style our high scores so they don't look
absurd on the page.

With the PHP, we get the top 10 scores, descending; this is how top
scores are traditionally displayed.  We can do different things with a
Score, so we add a method telling the script how to handle each call.
Finally, we output the scores in a JSON object.

In the JavaScript, we add a function to the view to get and display the
high scores.  The data isn't used in the game model, so we don't add a
similar function there.  We use a table to display the data, because the
data is tabular.  We only want to display the data associated with each
row, so we limit our display to array keys that contain a user.
Finally, we display our high scores after setting up the game.