PHP/JS - Store scores
We add some debugging code for viewing the return data of the PHP file
without it being attached to the game, and set up the score storing
functionality.

A Score is associated with a player (user) and a number representing the
score value.  We use sessions to prevent storing multiple scores for the
same game.  If we allowed that, one player could easily populate the
high score table by repeatedly saving a score.

We create a new Score, store it in the database, and return a JSON
object with this information.  A JSON object is returned simply as
verification that the score was stored successfully.

The #scoreBox reference is changed (a bug) to use the proper new naming
convention, and alt text is added to the images to keep our game
standards-compliant.

Finally, we attach an event handler to our "Save Score" button to save
the score.  This prompts the player for a user name, and uses POST
(because it's harder to manipulate than GET) to send data to the backend
for recording.  The High Scores table is refreshed with the new data in
case the score is a high score.